+++ src/webparts/columnFormatter/ColumnFormatterWebPart.ts
@@ -14,12 +14,13 @@ import pnp from "sp-pnp-js";
 import { Store, createStore } from 'redux';
 import { Provider, ProviderProps } from 'react-redux';
 import { IApplicationState } from './state/State';
-import { setContext } from './state/Actions';
+import { setContext, setHeight } from './state/Actions';
 import { cfReducer } from './state/Reducers';
 import { Environment, EnvironmentType } from '@microsoft/sp-core-library';
+import { PropertyFieldSpinButton } from '@pnp/spfx-property-controls/lib/PropertyFieldSpinButton';
 
 export interface IColumnFormatterWebPartProps {
-  description: string;
+  height: number;
 }
 
 export default class ColumnFormatterWebPart extends BaseClientSideWebPart<IColumnFormatterWebPartProps> {
@@ -29,7 +30,13 @@ export default class ColumnFormatterWebPart extends BaseClientSideWebPart<IColum
   public onInit(): Promise<void> {
 
     this.store = createStore(cfReducer);
-    this.store.dispatch(setContext(Environment.type !== EnvironmentType.Local, this.context.pageContext.web.absoluteUrl));
+    this.store.dispatch(
+      setContext(
+        Environment.type !== EnvironmentType.Local,
+        this.context.pageContext.web.absoluteUrl,
+        this.properties.height || 340
+      )
+    );
 
     return super.onInit().then(_ => {
       pnp.setup({
@@ -56,19 +63,31 @@ export default class ColumnFormatterWebPart extends BaseClientSideWebPart<IColum
     return Version.parse('1.0');
   }
 
+  public onPropertyHeightChanged(propertyPath: string, oldValue: any, newValue: any): void {
+    if(propertyPath == 'height' && oldValue !== newValue) {
+      this.store.dispatch(setHeight(Math.max(340, newValue)));
+    }
+    this.onPropertyPaneFieldChanged(propertyPath, oldValue, newValue);
+  }
+
   protected getPropertyPaneConfiguration(): IPropertyPaneConfiguration {
     return {
       pages: [
         {
-          header: {
-            description: strings.PropertyPaneDescription
-          },
           groups: [
             {
-              groupName: strings.BasicGroupName,
+              groupName: strings.PropertyBasicGroupName,
               groupFields: [
-                PropertyPaneTextField('description', {
-                  label: strings.DescriptionFieldLabel
+                PropertyFieldSpinButton('height', {
+                  label: strings.PropertyHeightLabel,
+                  initialValue: this.properties.height,
+                  onPropertyChange: this.onPropertyHeightChanged.bind(this),
+                  properties: this.properties,
+                  suffix: ' px',
+                  min: 340,
+                  step: 10,
+                  decimalPlaces: 0,
+                  key: 'height'
                 })
               ]
             }