+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Releases
 ## [1.1.0] - 2018-01-05
 ### Added
-- ?
+- Height property to property pane
 
 ### Changed
 - ?
 
 ### Removed
-- ?
+- Description property from property pane
 
 ## [1.0.0] - 2017-12-18
 ### Added